% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_scala}
\alias{as_scala}
\title{Converts an R object to an equivalent Scala object}
\usage{
as_scala(x, sc, ...)
}
\arguments{
\item{x}{R object coercible to a Scala object}

\item{sc}{A \code{spark_connection}}
}
\value{
a jobj reference to a Scala object
}
\description{
Converts an R object to an equivalent Scala object
}
