% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_scala_buffer}
\alias{as_scala_buffer}
\title{Converts a list or vector to a scalar buffer}
\usage{
as_scala_buffer(x, sc)
}
\arguments{
\item{x}{An R vector or list}

\item{sc}{A \code{spark_connection}}
}
\description{
Converts a list or vector to a scalar buffer
}
