% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_scala_some}
\alias{as_scala_some}
\title{Converts an object to a 'Some' object}
\usage{
as_scala_some(x, sc)
}
\arguments{
\item{x}{An R or jobj.}

\item{sc}{A \code{spark_connection}}
}
\description{
Converts an object to a 'Some' object
}
