% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{Attribute}
\alias{Attribute}
\alias{is.Attribute}
\alias{CategoricalAttribute}
\alias{is.CategoricalAttribute}
\title{Attribute specification}
\usage{
Attribute(simFn, distortionPrior)

is.Attribute(x)

CategoricalAttribute(distortionPrior)

is.CategoricalAttribute(x)
}
\arguments{
\item{simFn}{a \code{\link{SimilarityFn}}` object.}

\item{distortionPrior}{a \code{\link{BetaRV}} object. Specifies the prior
on the distortion probability for the attribute.}

\item{x}{an \code{R} object}
}
\value{
The \code{Attribute} constructor returns an \code{Attribute} object.

\code{is.Attribute} returns TRUE if the argument is an \code{Attribute}
object and FALSE otherwise.

The \code{CategoricalAttribute} constructor returns a
\code{CategoricalAttribute} object.
It is intended for modeling categorical attributes, and uses a
a constant similarity function.

\code{is.CategoricalAttribute} returns TRUE if the argument is a
\code{CategoricalAttribute} object and FALSE otherwise.
}
\description{
Represents an entity attribute that appears in the data: e.g. 'name of
company', 'age', 'zip code', etc.
}
\details{
An \code{Attribute} object contains all of the model parameters associated
with an entity attribute.
This includes:
\itemize{
\item the prior distribution over the distortion probability (likelihood
that an entity attribute is distorted in the data), and
\item the distance function (parameterizes the distortion distribution).
}

A named list of \code{Attribute} objects is required when initializing
the model (see \code{\link{initializeState}}).
}
