% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connect-fcns.R
\name{aws.connect}
\alias{aws.connect}
\title{aws.connect}
\usage{
aws.connect(
  usr,
  pw,
  db = "geoseg_1",
  driver = c("Postgres", "RPostgreSQL"),
  pool = FALSE,
  port = 5432
)
}
\description{
connects to my amazon web service (aws) database.
}
