% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{clusterSizeDistribution}
\alias{clusterSizeDistribution}
\alias{clusterSizeDistribution,linkagechain-method}
\alias{clusterSizeDistribution,dblinkresult-method}
\title{Cluster size distribution along the linkage chain}
\usage{
clusterSizeDistribution(x, ...)

\S4method{clusterSizeDistribution}{linkagechain}(x, ...)

\S4method{clusterSizeDistribution}{dblinkresult}(x, ...)
}
\arguments{
\item{x}{A \code{linkagechain} or \code{dblinkresult} object.}
}
\value{
A tibble representing the cluster size distribution along the chain.
}
\description{
TODO
}
\seealso{
\code{\link{partitionSizes}}, \code{\link{loadDiagnostics}}
}
