% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityFns.R
\name{ConstantSimFn}
\alias{ConstantSimFn}
\title{Constant similarity function}
\usage{
ConstantSimFn()
}
\value{
a \code{ConstantSimFn} object
}
\description{
Represents a similarity function that returns zero for all pairs of values
}
