% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{initializeState}
\alias{initializeState}
\title{Initialize the model state}
\usage{
initializeState(
  sc,
  data,
  attributeSpecs,
  recIdColname,
  partitioner,
  populationSize,
  fileIdColname = NULL,
  randomSeed = 1L,
  maxClusterSize = 10L
)
}
\arguments{
\item{sc}{A \code{spark_connection}}

\item{data}{A Spark DataFrame or an \R object that can be cast to a
Spark DataFrame}

\item{attributeSpecs}{A named list of \code{\link{Attribute}} objects. Each entry
in the list specifies the model parameters for an entity attribute,
and should match one of the column names (attributes) in
\code{data}.}

\item{recIdColname}{Column name in \code{data} that contains unique record
identifiers.}

\item{partitioner}{A \code{\linkS4class{Partitioner}} object which specifies how to
partition the space of entities (optional). If NULL, the entities are
not partitioned at all, however this can severely hinder scalability.}

\item{populationSize}{Specifies the prior on the size of the population of
latent entities. An integer specifies a fixed population size, a
a \code{\link{ShiftedNegBinomRV}} object specifies a shifted negative binomial
prior.}

\item{fileIdColname}{Column name in \code{data} that contains contains
file/source identifiers for the records. If NULL, the records are assumed
to be from a single file/source.}

\item{randomSeed}{An integer random seed.}

\item{maxClusterSize}{A guess at the maximum cluster size in \code{data}.}
}
\value{
A \code{state_jobj} object
}
\description{
Initialize the model state
}
\seealso{
\code{\link{loadState}}
}
