% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.scalar}
\alias{is.scalar}
\title{Checks whether the input is a scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{input}
}
\value{
TRUE if \code{x} is a scalar and FALSE otherwise
}
\description{
Checks whether the input is a scalar
}
