% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityFns.R
\name{LevenshteinSimFn}
\alias{LevenshteinSimFn}
\title{Levenshtein similarity function}
\usage{
LevenshteinSimFn(threshold, maxSimilarity)
}
\value{
a \code{LevenshteinSimFn} object
}
\description{
Levenshtein similarity function
}
