% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkageChain.R
\name{loadDBlinkResult}
\alias{loadDBlinkResult}
\title{Load a dblink result from disk}
\usage{
loadDBlinkResult(sc, projectPath)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{projectPath}{path to the project directory.}
}
\value{
A \code{LinkageChain} jobj.
}
\description{
Load a dblink result from disk
}
