% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkageChain.R
\name{loadLinkageChain}
\alias{loadLinkageChain}
\title{Load samples of the linkage structure}
\usage{
loadLinkageChain(sc, projectPath)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{projectPath}{path to the project directory.}
}
\value{
A \code{LinkageChain} jobj.
}
\description{
Load samples of the linkage structure
}
