% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{loadState}
\alias{loadState}
\title{Load a saved state from disk}
\usage{
loadState(sc, projectPath)
}
\arguments{
\item{projectPath}{Path to the project directory.}
}
\value{
A \code{state_jobj} object.
}
\description{
Load a saved state from disk
}
