% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{partitionSizes}
\alias{partitionSizes}
\alias{partitionSizes,linkagechain-method}
\alias{partitionSizes,dblinkresult-method}
\title{Partition sizes along the linkage chain}
\usage{
partitionSizes(x, ...)

\S4method{partitionSizes}{linkagechain}(x)

\S4method{partitionSizes}{dblinkresult}(x, ...)
}
\arguments{
\item{x}{A \code{linkagechain} or \code{dblinkresult} object.}
}
\value{
A tibble representing the partition sizes along the chain.
}
\description{
TODO
}
\seealso{
\code{\link{clusterSizeDistribution}}, \code{\link{loadDiagnostics}}
}
