% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-query-fcns.R
\name{persistent.query.division}
\alias{persistent.query.division}
\title{persistent.query.division}
\usage{
persistent.query.division(connection, region, div.name)
}
\description{
Applies query division until a non-null result is retrieved. A workaround to
an issue when the db is queried from within a shiny app. Not neccessary
outside of shiny context.
}
