% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connect-fcns.R
\name{princeton.db.connect}
\alias{princeton.db.connect}
\title{princeton.db.connect}
\usage{
princeton.db.connect(
  usr,
  pw,
  db = "shp",
  driver = c("Postgres", "RPostgreSQL"),
  pool = FALSE,
  port = 5432
)
}
\arguments{
\item{usr}{Username for connecting to the datbase.}

\item{pw}{password.}

\item{db}{name of database to connect to.}

\item{driver}{Driver to use. Parsed from text to RPostgres:Postgres() or
DBI::dbDriver("PostgreSQL"). First (default) option has been more
convenient for me.}

\item{pool}{Whether or not to open a connection as pool. Defaults to false.}
}
\description{
Stores code and some options to connect to princeton's postgres database.
Supply a username and password; default options should be appropriate
otherwise.
}
