% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-query-fcns.R
\name{q_db2sf}
\alias{q_db2sf}
\title{q_db2sf}
\usage{
q_db2sf(connection, schema_table)
}
\arguments{
\item{schema_table}{Schema and table in form "schema.table"}
}
\description{
Quick database to sf. Conveninece fcn to read spatial info from db into an sf
object in R.
}
