% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-query-fcns.R
\name{query.by.geoid}
\alias{query.by.geoid}
\title{query.by.geoid}
\usage{
query.by.geoid(con, region.type, geoid, tbl_name = "geopaired_hwys")
}
\description{
Both slower and less convenient than above, so not included.
}
