% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-query-fcns.R
\name{query.division}
\alias{query.division}
\title{query.division}
\usage{
query.division(connection, region, div.name, crs = 4326)
}
\arguments{
\item{connection}{a connection, as retrieved from \code{princeton.db.connect}}

\item{region}{A sf object}
}
\description{
Extracts all geometries from a postgis table that lay in the bounding box of
the supplied \code{region}.
}
