% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{sharedMostProbableClusters}
\alias{sharedMostProbableClusters}
\alias{sharedMostProbableClusters,linkagechain-method}
\alias{sharedMostProbableClusters,mostprobableclusters-method}
\alias{sharedMostProbableClusters,dblinkresult-method}
\title{Shared Most Probable Clusters}
\usage{
sharedMostProbableClusters(x, ...)

\S4method{sharedMostProbableClusters}{linkagechain}(x, ...)

\S4method{sharedMostProbableClusters}{mostprobableclusters}(x, ...)

\S4method{sharedMostProbableClusters}{dblinkresult}(x, ...)
}
\arguments{
\item{x}{A \code{dblinkresult} object as returned by \code{\link{runInference}} or a
\code{linkagechain} object as returned by \code{\link{loadLinkageChain}}.}

\item{m_jobj}{An optional \code{mostprobableclusters} object as returned by
\code{\link{mostProbableClusters}}. If provided, the function can skip
computing the most probable clusters.}
}
\value{
A \code{clusters} object.
}
\description{
Computes a point estimate of the most likely clustering that obeys
transitivity constraints based on posterior samples. The method was
introduced by Steorts et al. (2016), where it is referred to as the
method of \emph{shared most probable maximal matching sets}.
}
\references{
Steorts, R. C., Hall, R. & Fienberg, S. E. A Bayesian Approach
to Graphical Record Linkage and Deduplication. \emph{JASA} \strong{111},
1660–1672 (2016).
}
