% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityFns.R
\docType{class}
\name{SimFn-class}
\alias{SimFn-class}
\title{Similarity function}
\description{
Virtual class for a similarity function
}
\seealso{
Constructors for sub-classes include: \code{\link{ConstantSimFn}} and
\code{\link{LevenshteinSimFn}}.
}
