% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-convenience-fcns.R
\name{tbls.in.schema}
\alias{tbls.in.schema}
\title{tbls.in.schema}
\usage{
tbls.in.schema(con, schema)
}
\arguments{
\item{schema}{schema you want to check}
}
\description{
lists tables in schema. use \code{DBI::dbListObjects} to view different
schema.
}
