% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioners.R
\name{KDTreePartitioner}
\alias{KDTreePartitioner}
\title{k-d tree partitioner}
\usage{
KDTreePartitioner(numLevels, attributes)
}
\arguments{
\item{numLevels}{the depth/number of levels of the tree. The partitions are
the leaves of the tree, hence the number of partitions is given by
\code{2^numLevels}.}

\item{attributes}{splits are performed by cycling through the attributes
in this vector.}
}
\description{
k-d tree partitioner
}
