% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{mostProbableClusters}
\alias{mostProbableClusters}
\alias{mostProbableClusters,linkagechain-method}
\alias{mostProbableClusters,dblinkresult-method}
\title{Most Probable Clusters}
\usage{
mostProbableClusters(x, ...)

\S4method{mostProbableClusters}{linkagechain}(x, ...)

\S4method{mostProbableClusters}{dblinkresult}(x, ...)
}
\arguments{
\item{x}{a \code{dblinkresult} object as returned by \code{\link{runInference}}, or
a \code{linkagestructure} as returned by \code{\link{loadLinkageChain}}.}
}
\value{
A \code{mostprobableclusters} object.
}
\description{
Computes the most probable cluster for each record in the data set based on
posterior samples. The collection of most probable clusters is not
guaranteed to obey transitivity of closure. To obtain a transitive
clustering, one can apply the \code{\link{sharedMostProbableClusters}}
function to the output of this function. In (Steorts et al. 2016), the
most probable clusters are referred to as \emph{most probable maximal
matching sets}.
}
\references{
Steorts, R. C., Hall, R. & Fienberg, S. E. A Bayesian Approach
to Graphical Record Linkage and Deduplication. \emph{JASA} \strong{111},
1660–1672 (2016).
}
\seealso{
The \code{\link{sharedMostProbableClusters}} function computes a point estimate
from the most probable clusters (the output of this function), which
obeys transitivity constraints.
}
